more msvc fixes

+++ cpp/src/arrow/util/bit_util.h
@@ -112,8 +112,8 @@ static constexpr uint8_t kBytePopcount[] = {
     5, 4, 5, 5, 6, 3, 4, 4, 5, 4, 5, 5, 6, 4, 5, 5, 6, 5, 6, 6, 7, 3, 4, 4, 5, 4, 5, 5, 6,
     4, 5, 5, 6, 5, 6, 6, 7, 4, 5, 5, 6, 5, 6, 6, 7, 5, 6, 6, 7, 6, 7, 7, 8};
 
-static inline int PopCount(uint64_t bitmap) { return ARROW_POPCOUNT64(bitmap); }
-static inline int PopCount(uint32_t bitmap) { return ARROW_POPCOUNT32(bitmap); }
+static inline uint64_t PopCount(uint64_t bitmap) { return ARROW_POPCOUNT64(bitmap); }
+static inline uint32_t PopCount(uint32_t bitmap) { return ARROW_POPCOUNT32(bitmap); }
 
 //
 // Bit-related computations on integer values
